pyproject.toml
[project]
name = "dress-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama==0.4.6",
    "fastapi==0.128.0",
    "gitpython==3.1.46",
    "httpx>=0.28.1",
    "python-dotenv==1.2.1",
    "uvicorn==0.40.0",
]
